pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sweepstakes-agent"
version = "4.0.0"
description = "An AI agent that automatically discovers and enters free, legitimate sweepstakes using browser-use"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Daniel Sheils" },
]
keywords = ["sweepstakes", "automation", "browser-use", "ai-agent", "web-automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
]

dependencies = [
    "browser-use>=0.12.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0",
    "gradio>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]

[project.urls]
Homepage = "https://github.com/RhythrosaLabs/sweepstakes-agent"
Repository = "https://github.com/RhythrosaLabs/sweepstakes-agent"
Issues = "https://github.com/RhythrosaLabs/sweepstakes-agent/issues"

[project.scripts]
sweepstakes = "sweepstakes.__main__:cli_entry"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
